requirements.txt
apache-airflow==2.7.1
psycopg2==2.9.7
pandas==2.1.0